+++ Changelog.md
@@ -1,3 +1,7 @@
+# 11.0.5-20241121-1
+* Small refactoring of DialogModule.
+* Added option to send a paste from the chat window.
+
 # 11.0.5-20241023-1
 * Updated TOC to 11.0.5